Authenticate can now force re-authentication
This is actually a batch of changes by John Sutton [DAMTP], which he
describes as "enhancements and bug-fixes". The largest change he
describes thus:

"The main function authenticate($authassertionid = NULL, $testauthonly
= NULL) now takes the two arguments:

$authassertionid - is used to force re-authentication.  If specified,
it can be read as “authenticate(‘but this ticket will not do’)”.  I
use it to implement an idle timeout as follows: on every successful
authentication, retrieve the current ticket id using the id() function
and store it somewhere.  When you want to force re-authentication,
pass this ticket id back in as the first argument to authenticate().

$testauthonly - a boolean which if set true means to only test the
authentication state, i.e., return true or false accordingly but do
NOT send any headers.  This is useful when you are working with ajax
requests."